+++ clicktypes/network/__init__.py
@@ -1,4 +1,9 @@
 # -*- coding: utf-8 -*-
 
-# Import relevant classes
+# Import relevant classes for networking
+
+## IP Networks
 from .network import IPNetworkParam, IPv4NetworkParam, IPv6NetworkParam
+
+## IP Addresses
+from .address import IPAddressParam, IPv4AddressParam, IPv6AddressParam